requirements.txt
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gspread-formatting
requests
psycopg2-binary
openai
